C++ files split up into smaller files with needed headers for each
