+++ R/RcppExports.R
@@ -1,32 +1,16 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-QuickReg <- function(X, Y) {
-    .Call('_bdfm_QuickReg', PACKAGE = 'bdfm', X, Y)
+EstDFM <- function(B, Bp, Jb, lam_B, q, nu_q, H, Hp, lam_H, R, nu_r, Y, freq, LD, store_Y = FALSE, store_idx = 0L, reps = 1000L, burn = 500L, verbose = FALSE) {
+    .Call('_bdfm_EstDFM', PACKAGE = 'bdfm', B, Bp, Jb, lam_B, q, nu_q, H, Hp, lam_H, R, nu_r, Y, freq, LD, store_Y, store_idx, reps, burn, verbose)
 }
 
-UVreg <- function(x, y, rm_outlier = 0L) {
-    .Call('_bdfm_UVreg', PACKAGE = 'bdfm', x, y, rm_outlier)
+Ksmoother <- function(A, Q, HJ, R, Y) {
+    .Call('_bdfm_Ksmoother', PACKAGE = 'bdfm', A, Q, HJ, R, Y)
 }
 
-comp_form <- function(B) {
-    .Call('_bdfm_comp_form', PACKAGE = 'bdfm', B)
-}
-
-mvrnrm <- function(n, mu, Sigma) {
-    .Call('_bdfm_mvrnrm', PACKAGE = 'bdfm', n, mu, Sigma)
-}
-
-rinvwish <- function(n, v, S) {
-    .Call('_bdfm_rinvwish', PACKAGE = 'bdfm', n, v, S)
-}
-
-invchisq <- function(nu, scale) {
-    .Call('_bdfm_invchisq', PACKAGE = 'bdfm', nu, scale)
-}
-
-stack_obs <- function(nn, p, r = 0L) {
-    .Call('_bdfm_stack_obs', PACKAGE = 'bdfm', nn, p, r)
+KestExact <- function(A, Q, H, R, Y, itc, m, p) {
+    .Call('_bdfm_KestExact', PACKAGE = 'bdfm', A, Q, H, R, Y, itc, m, p)
 }
 
 J_MF <- function(days, m, ld, sA) {
@@ -61,19 +45,31 @@ Identify <- function(H, q) {
     .Call('_bdfm_Identify', PACKAGE = 'bdfm', H, q)
 }
 
-EstDFM <- function(B, Bp, Jb, lam_B, q, nu_q, H, Hp, lam_H, R, nu_r, Y, freq, LD, store_Y = FALSE, store_idx = 0L, reps = 1000L, burn = 500L, verbose = FALSE) {
-    .Call('_bdfm_EstDFM', PACKAGE = 'bdfm', B, Bp, Jb, lam_B, q, nu_q, H, Hp, lam_H, R, nu_r, Y, freq, LD, store_Y, store_idx, reps, burn, verbose)
+QuickReg <- function(X, Y) {
+    .Call('_bdfm_QuickReg', PACKAGE = 'bdfm', X, Y)
 }
 
-Ksmoother <- function(A, Q, HJ, R, Y) {
-    .Call('_bdfm_Ksmoother', PACKAGE = 'bdfm', A, Q, HJ, R, Y)
+UVreg <- function(x, y, rm_outlier = 0L) {
+    .Call('_bdfm_UVreg', PACKAGE = 'bdfm', x, y, rm_outlier)
 }
 
-KestExact <- function(A, Q, H, R, Y, itc, m, p) {
-    .Call('_bdfm_KestExact', PACKAGE = 'bdfm', A, Q, H, R, Y, itc, m, p)
+comp_form <- function(B) {
+    .Call('_bdfm_comp_form', PACKAGE = 'bdfm', B)
 }
 
-KSeas <- function(B, q, M, r, Y, N) {
-    .Call('_bdfm_KSeas', PACKAGE = 'bdfm', B, q, M, r, Y, N)
+mvrnrm <- function(n, mu, Sigma) {
+    .Call('_bdfm_mvrnrm', PACKAGE = 'bdfm', n, mu, Sigma)
+}
+
+rinvwish <- function(n, v, S) {
+    .Call('_bdfm_rinvwish', PACKAGE = 'bdfm', n, v, S)
+}
+
+invchisq <- function(nu, scale) {
+    .Call('_bdfm_invchisq', PACKAGE = 'bdfm', nu, scale)
+}
+
+stack_obs <- function(nn, p, r = 0L) {
+    .Call('_bdfm_stack_obs', PACKAGE = 'bdfm', nn, p, r)
 }
